requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.2.0
python-docx==1.1.0
PyPDF2==3.0.1 